Create ca.csv
Capture agent description file:
This is an excel-generated file to describe the capture agents
The "id" column makes the link to the agents-json. 

All other columns are shown in the info view, columns starting with "- " will not be used to filter Agents in the Utility view. 

The two lines in the beginning starting with -lang-en and -lang-es are for the localization of the enrich data. They are translating the Headings.